+++ RELEASES.md
@@ -16,6 +16,7 @@
 - the `Pool` trait has been split in two, with the regeneration-specific mechanics handled in `RegeneratingPool`, to make the construction of non-regenerating pools much more intuitive
 - added the `Pool::is_empty` and `Pool::is_full` helper methods to the `Pool` trait
 - added `Add`, `Sub`, `AddAssign` and `SubAssign` implementations to the premade `Life` and `Mana` types and their corresponding pools
+- added the `Display` trait to `Life`, `Mana`, `LifePool` and `ManaPool`
 
 ## Version 0.5